+++ pkg/server/server.go
@@ -36,6 +36,7 @@ import (
 	"github.com/Altinity/clickhouse-backup/v2/pkg/server/metrics"
 	"github.com/Altinity/clickhouse-backup/v2/pkg/status"
 	"github.com/Altinity/clickhouse-backup/v2/pkg/utils"
+	"github.com/google/uuid"
 )
 
 type APIServer struct {
@@ -870,10 +871,14 @@ func (api *APIServer) httpCreateHandler(w http.ResponseWriter, r *http.Request) 
 	backupName := backup.NewBackupName()
 	schemaOnly := false
 	createRBAC := false
+	rbacOnly := false
 	createConfigs := false
+	configsOnly := false
 	checkPartsColumns := true
 	fullCommand := "create"
 	query := r.URL.Query()
+	operationId, _ := uuid.NewUUID()
+
 	if tp, exist := query["table"]; exist {
 		tablePattern = tp[0]
 		fullCommand = fmt.Sprintf("%s --tables=\"%s\"", fullCommand, tablePattern)
@@ -885,23 +890,25 @@ func (api *APIServer) httpCreateHandler(w http.ResponseWriter, r *http.Request) 
 		partitionsToBackup = append(partitionsToBackup, partitions...)
 		fullCommand = fmt.Sprintf("%s --partitions=\"%s\"", fullCommand, strings.Join(partitions, "\" --partitions=\""))
 	}
-	if schema, exist := query["schema"]; exist {
-		schemaOnly, _ = strconv.ParseBool(schema[0])
-		if schemaOnly {
-			fullCommand = fmt.Sprintf("%s --schema", fullCommand)
-		}
+	if _, exist := query["schema"]; exist {
+		schemaOnly = true
+		fullCommand += " --schema"
 	}
-	if rbac, exist := query["rbac"]; exist {
-		createRBAC, _ = strconv.ParseBool(rbac[0])
-		if createRBAC {
-			fullCommand = fmt.Sprintf("%s --rbac", fullCommand)
-		}
+	if _, exist := query["rbac"]; exist {
+		createRBAC = true
+		fullCommand += " --rbac"
 	}
-	if configs, exist := query["configs"]; exist {
-		createConfigs, _ = strconv.ParseBool(configs[0])
-		if createConfigs {
-			fullCommand = fmt.Sprintf("%s --configs", fullCommand)
-		}
+	if _, exist := query["rbac-only"]; exist {
+		rbacOnly = true
+		fullCommand += " --rbac-only"
+	}
+	if _, exist := query["configs"]; exist {
+		createConfigs = true
+		fullCommand += " --configs"
+	}
+	if _, exist := query["configs-only"]; exist {
+		configsOnly = true
+		fullCommand += " --configs-only"
 	}
 
 	if partsColumns, exist := query["check_parts_columns"]; exist {
@@ -925,12 +932,12 @@ func (api *APIServer) httpCreateHandler(w http.ResponseWriter, r *http.Request) 
 	go func() {
 		err, _ := api.metrics.ExecuteWithMetrics("create", 0, func() error {
 			b := backup.NewBackuper(cfg)
-			return b.CreateBackup(backupName, diffFromRemote, tablePattern, partitionsToBackup, schemaOnly, createRBAC, false, createConfigs, false, checkPartsColumns, api.clickhouseBackupVersion, commandId)
+			return b.CreateBackup(backupName, diffFromRemote, tablePattern, partitionsToBackup, schemaOnly, createRBAC, rbacOnly, createConfigs, configsOnly, checkPartsColumns, api.clickhouseBackupVersion, commandId)
 		})
 		if err != nil {
 			log.Error().Msgf("API /backup/create error: %v", err)
 			status.Current.Stop(commandId, err)
-			api.errorCallback(context.Background(), err, callback)
+			api.errorCallback(context.Background(), err, operationId.String(), callback)
 			return
 		}
 		go func() {
@@ -940,16 +947,18 @@ func (api *APIServer) httpCreateHandler(w http.ResponseWriter, r *http.Request) 
 		}()
 
 		status.Current.Stop(commandId, nil)
-		api.successCallback(context.Background(), callback)
+		api.successCallback(context.Background(), operationId.String(), callback)
 	}()
 	api.sendJSONEachRow(w, http.StatusCreated, struct {
-		Status     string `json:"status"`
-		Operation  string `json:"operation"`
-		BackupName string `json:"backup_name"`
+		Status      string `json:"status"`
+		Operation   string `json:"operation"`
+		BackupName  string `json:"backup_name"`
+		OperationId string `json:"operation_id"`
 	}{
-		Status:     "acknowledged",
-		Operation:  "create",
-		BackupName: backupName,
+		Status:      "acknowledged",
+		Operation:   "create",
+		BackupName:  backupName,
+		OperationId: operationId.String(),
 	})
 }
 
@@ -1118,6 +1127,7 @@ func (api *APIServer) httpUploadHandler(w http.ResponseWriter, r *http.Request) 
 	schemaOnly := false
 	resume := false
 	fullCommand := "upload"
+	operationId, _ := uuid.NewUUID()
 
 	if _, exist := query["delete-source"]; exist {
 		deleteSource = true
@@ -1167,7 +1177,7 @@ func (api *APIServer) httpUploadHandler(w http.ResponseWriter, r *http.Request) 
 		if err != nil {
 			log.Error().Msgf("Upload error: %v", err)
 			status.Current.Stop(commandId, err)
-			api.errorCallback(context.Background(), err, callback)
+			api.errorCallback(context.Background(), err, operationId.String(), callback)
 			return
 		}
 		go func() {
@@ -1176,20 +1186,22 @@ func (api *APIServer) httpUploadHandler(w http.ResponseWriter, r *http.Request) 
 			}
 		}()
 		status.Current.Stop(commandId, nil)
-		api.successCallback(context.Background(), callback)
+		api.successCallback(context.Background(), operationId.String(), callback)
 	}()
 	api.sendJSONEachRow(w, http.StatusOK, struct {
-		Status     string `json:"status"`
-		Operation  string `json:"operation"`
-		BackupName string `json:"backup_name"`
-		BackupFrom string `json:"backup_from,omitempty"`
-		Diff       bool   `json:"diff"`
+		Status      string `json:"status"`
+		Operation   string `json:"operation"`
+		BackupName  string `json:"backup_name"`
+		BackupFrom  string `json:"backup_from,omitempty"`
+		Diff        bool   `json:"diff"`
+		OperationId string `json:"operation_id"`
 	}{
-		Status:     "acknowledged",
-		Operation:  "upload",
-		BackupName: name,
-		BackupFrom: diffFrom,
-		Diff:       diffFrom != "",
+		Status:      "acknowledged",
+		Operation:   "upload",
+		BackupName:  name,
+		BackupFrom:  diffFrom,
+		Diff:        diffFrom != "",
+		OperationId: operationId.String(),
 	})
 }
 
@@ -1217,8 +1229,11 @@ func (api *APIServer) httpRestoreHandler(w http.ResponseWriter, r *http.Request)
 	dropExists := false
 	ignoreDependencies := false
 	restoreRBAC := false
+	rbacOnly := false
 	restoreConfigs := false
+	configsOnly := false
 	fullCommand := "restore"
+	operationId, _ := uuid.NewUUID()
 
 	query := r.URL.Query()
 	if tp, exist := query["table"]; exist {
@@ -1286,9 +1301,17 @@ func (api *APIServer) httpRestoreHandler(w http.ResponseWriter, r *http.Request)
 		restoreRBAC = true
 		fullCommand += " --rbac"
 	}
+	if _, exist := query["rbac-only"]; exist {
+		rbacOnly = true
+		fullCommand += " --rbac-only"
+	}
 	if _, exist := query["configs"]; exist {
 		restoreConfigs = true
 		fullCommand += " --configs"
+	}
+	if _, exist := query["configs"]; exist {
+		configsOnly = true
+		fullCommand += " --configs-only"
 	}
 
 	name := utils.CleanBackupNameRE.ReplaceAllString(vars["name"], "")
@@ -1305,24 +1328,26 @@ func (api *APIServer) httpRestoreHandler(w http.ResponseWriter, r *http.Request)
 	go func() {
 		err, _ := api.metrics.ExecuteWithMetrics("restore", 0, func() error {
 			b := backup.NewBackuper(api.config)
-			return b.Restore(name, tablePattern, databaseMappingToRestore, tableMappingToRestore, partitionsToBackup, schemaOnly, dataOnly, dropExists, ignoreDependencies, restoreRBAC, false, restoreConfigs, false, api.cliApp.Version, commandId)
+			return b.Restore(name, tablePattern, databaseMappingToRestore, tableMappingToRestore, partitionsToBackup, schemaOnly, dataOnly, dropExists, ignoreDependencies, restoreRBAC, rbacOnly, restoreConfigs, configsOnly, api.cliApp.Version, commandId)
 		})
 		status.Current.Stop(commandId, err)
 		if err != nil {
 			log.Error().Msgf("API /backup/restore error: %v", err)
-			api.errorCallback(context.Background(), err, callback)
+			api.errorCallback(context.Background(), err, operationId.String(), callback)
 			return
 		}
-		api.successCallback(context.Background(), callback)
+		api.successCallback(context.Background(), operationId.String(), callback)
 	}()
 	api.sendJSONEachRow(w, http.StatusOK, struct {
-		Status     string `json:"status"`
-		Operation  string `json:"operation"`
-		BackupName string `json:"backup_name"`
+		Status      string `json:"status"`
+		Operation   string `json:"operation"`
+		BackupName  string `json:"backup_name"`
+		OperationId string `json:"operation_id"`
 	}{
-		Status:     "acknowledged",
-		Operation:  "restore",
-		BackupName: name,
+		Status:      "acknowledged",
+		Operation:   "restore",
+		BackupName:  name,
+		OperationId: operationId.String(),
 	})
 }
 
@@ -1346,6 +1371,7 @@ func (api *APIServer) httpDownloadHandler(w http.ResponseWriter, r *http.Request
 	schemaOnly := false
 	resume := false
 	fullCommand := "download"
+	operationId, _ := uuid.NewUUID()
 
 	if tp, exist := query["table"]; exist {
 		tablePattern = tp[0]
@@ -1381,7 +1407,7 @@ func (api *APIServer) httpDownloadHandler(w http.ResponseWriter, r *http.Request
 		if err != nil {
 			log.Error().Msgf("API /backup/download error: %v", err)
 			status.Current.Stop(commandId, err)
-			api.errorCallback(context.Background(), err, callback)
+			api.errorCallback(context.Background(), err, operationId.String(), callback)
 			return
 		}
 		go func() {
@@ -1390,16 +1416,18 @@ func (api *APIServer) httpDownloadHandler(w http.ResponseWriter, r *http.Request
 			}
 		}()
 		status.Current.Stop(commandId, nil)
-		api.successCallback(context.Background(), callback)
+		api.successCallback(context.Background(), operationId.String(), callback)
 	}()
 	api.sendJSONEachRow(w, http.StatusOK, struct {
-		Status     string `json:"status"`
-		Operation  string `json:"operation"`
-		BackupName string `json:"backup_name"`
+		Status      string `json:"status"`
+		Operation   string `json:"operation"`
+		BackupName  string `json:"backup_name"`
+		OperationId string `json:"operation_id"`
 	}{
-		Status:     "acknowledged",
-		Operation:  "download",
-		BackupName: name,
+		Status:      "acknowledged",
+		Operation:   "download",
+		BackupName:  name,
+		OperationId: operationId.String(),
 	})
 }